+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerTablePanel.java
@@ -69,7 +69,7 @@ public final class CustomerTablePanel extends EntityTablePanel {
 
 	private JasperPrint fillCustomerReport() throws DatabaseException, ReportException {
 		Collection<Long> customerIDs = Entity.values(Customer.ID,
-						tableModel().selectionModel().getSelectedItems());
+						tableModel().selectionModel().selectedItems());
 		Map<String, Object> reportParameters = new HashMap<>();
 		reportParameters.put("CUSTOMER_IDS", customerIDs);