+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -655,7 +655,7 @@ public class EntityTablePanel extends JPanel {
 		requireNonNull(attributeToEdit);
 		if (!tableModel.selectionModel().isSelectionEmpty()) {
 			editDialogBuilder(attributeToEdit)
-							.edit(tableModel.selectionModel().getSelectedItems());
+							.edit(tableModel.selectionModel().selectedItems());
 		}
 	}
 
@@ -664,7 +664,7 @@ public class EntityTablePanel extends JPanel {
 	 */
 	public final void viewDependencies() {
 		if (!tableModel.selectionModel().isSelectionEmpty()) {
-			displayDependenciesDialog(tableModel.selectionModel().getSelectedItems(), tableModel.connectionProvider(), this);
+			displayDependenciesDialog(tableModel.selectionModel().selectedItems(), tableModel.connectionProvider(), this);
 		}
 	}
 
@@ -943,7 +943,7 @@ public class EntityTablePanel extends JPanel {
 	protected void onReferentialIntegrityException(ReferentialIntegrityException exception) {
 		requireNonNull(exception);
 		if (configuration.referentialIntegrityErrorHandling == ReferentialIntegrityErrorHandling.DISPLAY_DEPENDENCIES) {
-			displayDependenciesDialog(tableModel.selectionModel().getSelectedItems(), tableModel.connectionProvider(),
+			displayDependenciesDialog(tableModel.selectionModel().selectedItems(), tableModel.connectionProvider(),
 							this, EDIT_PANEL_MESSAGES.getString("unknown_dependent_records"));
 		}
 		else {
@@ -1850,7 +1850,7 @@ public class EntityTablePanel extends JPanel {
 		@Override
 		public void execute() {
 			if (confirmDelete()) {
-				List<Entity> selectedItems = tableModel().selectionModel().getSelectedItems();
+				List<Entity> selectedItems = tableModel().selectionModel().selectedItems();
 				progressWorkerDialog(tableModel().editModel().createDelete(selectedItems).prepare()::perform)
 								.title(EDIT_PANEL_MESSAGES.getString("deleting"))
 								.owner(EntityTablePanel.this)