+++ tools/loadtest/model/src/main/java/is/codion/tools/loadtest/model/DefaultLoadTestModel.java
@@ -124,7 +124,7 @@ final class DefaultLoadTestModel<T> implements LoadTestModel<T> {
 
 	@Override
 	public void removeSelectedApplications() {
-		applicationTableModel.selectionModel().getSelectedItems().stream()
+		applicationTableModel.selectionModel().selectedItems().stream()
 						.map(DefaultApplicationRow.class::cast)
 						.forEach(application -> loadTest.stop(application.applicationRunner));
 	}